Verify that messages are between 86 and 255 characters long
Without any padding at all, messages that are shorter than this won't be
greater than `n` when exponentiated, so the same message will always
"encrypt" to the same value regardless of the value of the key.

Messages longer than 255 characters will usually be greater than 2*`n` when
exponentiated, and will decrypt incorrectly.

All of this is because my implementation is (consciously) terrible at
the moment.